fix format

+++ contracts/Marketplace/Marketplace.sol
@@ -10,7 +10,6 @@ import "../PaymentRouter/IPaymentRouter.sol";
 import "../Tokens/IPazariTokenMVP.sol";
 
 contract AccessControlMP {
-
   // Maps admin addresses to bool
   mapping(address => bool) public isAdmin;
 
@@ -24,51 +23,46 @@ contract AccessControlMP {
   mapping(uint256 => address) public itemCreator;
 
   // Fires when admins are added or removed
-  event AdminAdded(
-    address indexed newAdmin, 
-    address indexed adminAuthorized, 
-    string memo, 
-    uint256 timestamp
-  );
+  event AdminAdded(address indexed newAdmin, address indexed adminAuthorized, string memo, uint256 timestamp);
   event AdminRemoved(
-    address indexed oldAdmin, 
-    address indexed adminAuthorized, 
-    string memo, 
+    address indexed oldAdmin,
+    address indexed adminAuthorized,
+    string memo,
     uint256 timestamp
   );
 
   // Fires when item admins are added or removed
   event ItemAdminAdded(
-    uint256 indexed itemID, 
-    address indexed newAdmin, 
-    address indexed adminAuthorized, 
-    string memo, 
+    uint256 indexed itemID,
+    address indexed newAdmin,
+    address indexed adminAuthorized,
+    string memo,
     uint256 timestamp
   );
   event ItemAdminRemoved(
-    uint256 indexed itemID, 
-    address indexed oldAdmin, 
-    address indexed adminAuthorized, 
-    string memo, 
+    uint256 indexed itemID,
+    address indexed oldAdmin,
+    address indexed adminAuthorized,
+    string memo,
     uint256 timestamp
   );
 
   // Fires when an address is blacklisted/whitelisted from the Pazari Marketplace
   event AddressBlacklisted(
-    address blacklistedAddress, 
-    address indexed adminAddress, 
-    string memo, 
+    address blacklistedAddress,
+    address indexed adminAddress,
+    string memo,
     uint256 timestamp
   );
   event AddressWhitelisted(
-    address whitelistedAddress, 
-    address indexed adminAddress, 
-    string memo, 
+    address whitelistedAddress,
+    address indexed adminAddress,
+    string memo,
     uint256 timestamp
   );
 
   constructor(address[] memory _adminAddresses) {
-    for(uint i = 0; i < _adminAddresses.length; i++){
+    for (uint256 i = 0; i < _adminAddresses.length; i++) {
       isAdmin[_adminAddresses[i]] = true;
     }
   }
@@ -78,59 +72,66 @@ contract AccessControlMP {
    * for everything else. See PaymentRouter for more details.
    */
   function _msgSender() public view returns (address) {
-    if(tx.origin != msg.sender && isAdmin[msg.sender]){
+    if (tx.origin != msg.sender && isAdmin[msg.sender]) {
       return tx.origin;
-    }
-    else return msg.sender;
+    } else return msg.sender;
   }
-  
+
   // Adds an address to isAdmin mapping
   // Emits AdminAdded event
   function addAdmin(address _newAddress, string memory _memo) external onlyAdmin returns (bool) {
     require(!isAdmin[_newAddress], "Address is already an admin");
-    
+
     isAdmin[_newAddress] = true;
-    
+
     emit AdminAdded(_newAddress, tx.origin, _memo, block.timestamp);
     return true;
   }
 
   // Adds an address to isItemAdmin mapping
   // Emits ItemAdminAdded event
-  function addItemAdmin(uint256 _itemID, address _newAddress, string memory _memo) external onlyItemAdmin(_itemID) returns (bool) {
-    require(isItemAdmin[_itemID][msg.sender] && isItemAdmin[_itemID][tx.origin], "Caller is not admin");    
-    require(!isItemAdmin[_itemID][_newAddress], "Address is already an item admin");    
+  function addItemAdmin(
+    uint256 _itemID,
+    address _newAddress,
+    string memory _memo
+  ) external onlyItemAdmin(_itemID) returns (bool) {
+    require(isItemAdmin[_itemID][msg.sender] && isItemAdmin[_itemID][tx.origin], "Caller is not admin");
+    require(!isItemAdmin[_itemID][_newAddress], "Address is already an item admin");
 
     isItemAdmin[_itemID][_newAddress] = true;
 
     emit ItemAdminAdded(_itemID, _newAddress, _msgSender(), _memo, block.timestamp);
-    return true;   
+    return true;
   }
 
   // Removes an address from isAdmin mapping
   // Emits AdminRemoved event
   function removeAdmin(address _oldAddress, string memory _memo) external onlyAdmin returns (bool) {
     require(isAdmin[_oldAddress], "Address is not an admin");
-    
+
     isAdmin[_oldAddress] = false;
-    
+
     emit AdminRemoved(_oldAddress, tx.origin, _memo, block.timestamp);
     return true;
   }
 
   // Removes an address from isItemAdmin mapping
   // Emits ItemAdminRemoved event
-  function removeItemAdmin(uint256 _itemID, address _oldAddress, string memory _memo) external onlyItemAdmin(_itemID) returns (bool) {
-    require(isItemAdmin[_itemID][msg.sender] && isItemAdmin[_itemID][tx.origin], "Caller is not admin");    
-    require(isItemAdmin[_itemID][_oldAddress], "Address is not an admin");    
+  function removeItemAdmin(
+    uint256 _itemID,
+    address _oldAddress,
+    string memory _memo
+  ) external onlyItemAdmin(_itemID) returns (bool) {
+    require(isItemAdmin[_itemID][msg.sender] && isItemAdmin[_itemID][tx.origin], "Caller is not admin");
+    require(isItemAdmin[_itemID][_oldAddress], "Address is not an admin");
     require(itemCreator[_itemID] == _msgSender(), "Cannot remove item creator");
 
     isItemAdmin[_itemID][_oldAddress] = false;
 
     emit ItemAdminRemoved(_itemID, _oldAddress, _msgSender(), _memo, block.timestamp);
     return true;
-  }  
-  
+  }
+
   /**
    * @notice Toggles isBlacklisted for an address. Can only be called by Pazari
    * Marketplace admins. Other contracts that implement address blacklisting
@@ -142,22 +143,21 @@ contract AccessControlMP {
    * @dev Emits AddressBlacklisted event when _userAddress is blacklisted
    * @dev Emits AddressWhitelisted event when _userAddress is whitelisted
    */
-  function toggleBlacklist(address _userAddress, string memory _memo) external returns (bool){
+  function toggleBlacklist(address _userAddress, string memory _memo) external returns (bool) {
     require(isAdmin[msg.sender] && isAdmin[tx.origin], "Only Pazari admin");
     require(!isAdmin[_userAddress], "Cannot blacklist admins");
 
-    if(!isBlacklisted[_userAddress]){
+    if (!isBlacklisted[_userAddress]) {
       isBlacklisted[_userAddress] = true;
       emit AddressBlacklisted(_userAddress, _msgSender(), _memo, block.timestamp);
-    }
-    else {
+    } else {
       isBlacklisted[_userAddress] = false;
       emit AddressWhitelisted(_userAddress, _msgSender(), _memo, block.timestamp);
     }
 
     return true;
   }
- 
+
   /**
    * @notice Requires that both msg.sender and tx.origin be admins. This restricts all
    * calls to only Pazari-owned admin addresses, including wallets and contracts, and
@@ -166,7 +166,7 @@ contract AccessControlMP {
   modifier onlyAdmin() {
     require(isAdmin[msg.sender] && isAdmin[tx.origin], "Only Pazari-owned addresses");
     _;
-  } 
+  }
 
   modifier noBlacklist() {
     require(!isBlacklisted[_msgSender()], "Caller cannot be blacklisted");
@@ -176,14 +176,11 @@ contract AccessControlMP {
   // Restricts access to admins of a MarketItem
   modifier onlyItemAdmin(uint256 _itemID) {
     require(
-      itemCreator[_itemID] == _msgSender() || 
-      isItemAdmin[_itemID][_msgSender()] ||
-      isAdmin[_msgSender()],
+      itemCreator[_itemID] == _msgSender() || isItemAdmin[_itemID][_msgSender()] || isAdmin[_msgSender()],
       "Caller is neither admin nor item creator"
     );
     _;
   }
-
 }
 
 contract Marketplace is ERC1155Holder, AccessControlMP {
@@ -251,7 +248,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     bool forSale;
     uint256 itemLimit;
   }
-  
+
   // Counter for items with forSale == false
   Counters.Counter private itemsSoldOut;
 
@@ -394,7 +391,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     // Check that Marketplace's internal balance matches the token's balanceOf() value
     MarketItem memory item = marketItems[itemID - 1];
     assert(IERC1155(item.tokenContract).balanceOf(address(this), item.tokenID) == item.amount);
-  }  
+  }
 
   /**
    * @dev Private function that updates internal variables and storage for a new MarketItem
@@ -444,7 +441,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     itemCreator[itemID] = _msgSender();
     // Pushes MarketItem to marketItems[]
     marketItems.push(item);
-  
+
     // Push itemID to sellersMarketItems mapping array
     // _msgSender == sellerAddress
     sellersMarketItems[_msgSender()].push(item.itemID);
@@ -454,15 +451,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
 
     // Emits MarketItemCreated event
     // _msgSender == sellerAddress
-    emit MarketItemCreated(
-      itemID,
-      _tokenContract,
-      _tokenID,
-      _msgSender(),
-      _price,
-      _amount,
-      _paymentContract
-    );
+    emit MarketItemCreated(itemID, _tokenContract, _tokenID, _msgSender(), _price, _amount, _paymentContract);
   }
 
   /**
@@ -476,7 +465,12 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
    *
    * note Providing _amount == 0 will purchase the item's full itemLimit.
    */
-  function buyMarketItem(uint256 _itemID, uint256 _amount) external noBlacklist itemExists(_itemID) returns (bool) {
+  function buyMarketItem(uint256 _itemID, uint256 _amount)
+    external
+    noBlacklist
+    itemExists(_itemID)
+    returns (bool)
+  {
     // Pull data from itemID's MarketItem struct
     MarketItem memory item = marketItems[_itemID - 1];
     uint256 balance = IERC1155(item.tokenContract).balanceOf(_msgSender(), item.tokenID);
@@ -487,8 +481,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
 
     /* ========== CHECKS ========== */
     require(
-      !isItemAdmin[item.itemID][_msgSender()] ||
-      itemCreator[item.itemID] != _msgSender(), 
+      !isItemAdmin[item.itemID][_msgSender()] || itemCreator[item.itemID] != _msgSender(),
       "Can't buy your own item"
     );
     require(item.amount > 0, "Item sold out");
@@ -496,7 +489,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     require(balance < itemLimit, "Buyer already owns the item limit");
     // If _amount == 0, then purchase itemLimit - balance
     // If _amount + balance surpasses itemLimit, then purchase itemLimit - balance
-    if(_amount == 0 || _amount + balance > itemLimit){
+    if (_amount == 0 || _amount + balance > itemLimit) {
       _amount = itemLimit - balance;
     }
 
@@ -518,10 +511,16 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     require(IERC20(item.paymentContract).approve(address(this), totalCost), "ERC20 approval failure");
 
     // Pull payment tokens from msg.sender to Marketplace
-    require(IERC20(item.paymentContract).transferFrom(_msgSender(), address(this), totalCost), "ERC20 transfer failure");
+    require(
+      IERC20(item.paymentContract).transferFrom(_msgSender(), address(this), totalCost),
+      "ERC20 transfer failure"
+    );
 
     // Approve payment tokens for transfer to PaymentRouter
-    require(IERC20(item.paymentContract).approve(address(iPaymentRouter), totalCost), "ERC20 approval failure");
+    require(
+      IERC20(item.paymentContract).approve(address(iPaymentRouter), totalCost),
+      "ERC20 approval failure"
+    );
 
     // Send ERC20 tokens through PaymentRouter, isPush determines which function is used
     // note PaymentRouter functions make external calls to ERC20 contracts, thus they are interactions
@@ -545,14 +544,23 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
    *
    * @dev Emits ItemRestocked event
    */
-  function restockItem(uint256 _itemID, uint256 _amount) external noBlacklist onlyItemAdmin(_itemID) itemExists(_itemID) returns (bool) {
+  function restockItem(uint256 _itemID, uint256 _amount)
+    external
+    noBlacklist
+    onlyItemAdmin(_itemID)
+    itemExists(_itemID)
+    returns (bool)
+  {
     /* ========== CHECKS ========== */
     MarketItem memory item = marketItems[_itemID - 1];
-    require(IERC1155(item.tokenContract).balanceOf(_msgSender(), item.tokenID) >= _amount, "Insufficient token balance");
+    require(
+      IERC1155(item.tokenContract).balanceOf(_msgSender(), item.tokenID) >= _amount,
+      "Insufficient token balance"
+    );
 
     /* ========== EFFECTS ========== */
     // If item is out of stock
-    if(item.amount == 0){
+    if (item.amount == 0) {
       itemsSoldOut.decrement();
       item.forSale = true;
     }
@@ -563,7 +571,9 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     /* ========== INTERACTIONS ========== */
     IERC1155(item.tokenContract).safeTransferFrom(_msgSender(), address(this), item.tokenID, _amount, "");
 
-    assert(IERC1155(item.tokenContract).balanceOf(address(this), item.tokenID) == marketItems[_itemID - 1].amount);
+    assert(
+      IERC1155(item.tokenContract).balanceOf(address(this), item.tokenID) == marketItems[_itemID - 1].amount
+    );
     emit ItemRestocked(_itemID, _amount);
     return true;
   }
@@ -576,7 +586,13 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
    *
    * @dev Emits ItemPulled event
    */
-  function pullStock(uint256 _itemID, uint256 _amount) external noBlacklist onlyItemAdmin(_itemID) itemExists(_itemID) returns (bool) {
+  function pullStock(uint256 _itemID, uint256 _amount)
+    external
+    noBlacklist
+    onlyItemAdmin(_itemID)
+    itemExists(_itemID)
+    returns (bool)
+  {
     /* ========== CHECKS ========== */
     // Store initial values
     MarketItem memory item = marketItems[_itemID - 1];
@@ -672,7 +688,13 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
    *
    * @dev Emits ForSaleToggled event
    */
-  function toggleForSale(uint256 _itemID) external noBlacklist onlyItemAdmin(_itemID) itemExists(_itemID) returns (bool forSale) {
+  function toggleForSale(uint256 _itemID)
+    external
+    noBlacklist
+    onlyItemAdmin(_itemID)
+    itemExists(_itemID)
+    returns (bool forSale)
+  {
     // Create singleton array for _itemID
     uint256[] memory itemID = new uint256[](1);
     itemID[0] = _itemID;
@@ -703,9 +725,9 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
     MarketItem memory item = marketItems[_itemID - 1];
     // Caller must either be item's creator or a Pazari admin, no itemAdmins allowed
     require(
-      _msgSender() == itemCreator[_itemID] ||
-      isAdmin[_msgSender()],
-      "Only item creators and Pazari admins can delete items");
+      _msgSender() == itemCreator[_itemID] || isAdmin[_msgSender()],
+      "Only item creators and Pazari admins can delete items"
+    );
     // Require item has been completely unstocked and deactivated
     require(!item.forSale, "Deactivate item before deleting");
     require(item.amount == 0, "Pull all stock before deleting");
@@ -816,7 +838,7 @@ contract Marketplace is ERC1155Holder, AccessControlMP {
    * @notice Returns an array of MarketItems created the seller's address
    * @dev No restrictions for calling this
    */
-  function getSellersMarketItems(address _sellerAddress) public view returns (uint256[] memory){
+  function getSellersMarketItems(address _sellerAddress) public view returns (uint256[] memory) {
     return sellersMarketItems[_sellerAddress];
   }
